refactor(SEPs): rename SEP(Scientific evaluation panel) to Fap(Facility access panel)
