+++ example.custom-proposal-body.hbs
@@ -289,10 +289,10 @@
           </div>
         </div>
 
-        {{#if sepReviews}} {{#each sepReviews}}
-        <div class="row mt-4 my-3" data-book-index="SEP Reviews">
+        {{#if fapReviews}} {{#each fapReviews}}
+        <div class="row mt-4 my-3" data-book-index="Fap Reviews">
           <div class="col">
-            <h4 class="bold" class="mb-4">SEP REVIEW {{ $sum @index 1 }}</h4>
+            <h4 class="bold" class="mb-4">Fap Review {{ $sum @index 1 }}</h4>
 
             <div class="row my-2">
               <div class="col-4 bold">Grade:</div>
