+++ src/__tests__/fixtures/pdf-payloads.json
@@ -209,7 +209,7 @@
           ]
         }
       ],
-      "sepReviews": []
+      "fapReviews": []
     },
     {
       "samples": [
@@ -295,7 +295,7 @@
         }
       ],
       "genericTemplates": [],
-      "sepReviews": []
+      "fapReviews": []
     }
   ],
   "sample_test_1": [
