+++ src/__tests__/fixtures/xlsx-payloads.json
@@ -1,5 +1,5 @@
 {
-  "sep_test_1": {
+  "fap_test_1": {
     "data": [
       {
         "sheetName": "Instrument 1",
